ingestor/requirements.txt
llama-index
qdrant-client
requests
openai
python-dotenv